pyproject.toml
[project]
name = "face-access-control"
version = "0.1.0"
description = "A production-ready computer vision system for face-based access control"
requires-python = "==3.13.*"
dependencies = [
    "insightface>=0.7.3",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "onnxruntime>=1.16.0",
    "marimo>=0.4.0",
]

[tool.marimo.runtime]
output_max_bytes = 20_000_000
